backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
python-dotenv==1.0.1

# RAG & IR
llama-index-core==0.11.0
llama-index-embeddings-huggingface==0.3.0
sentence-transformers==3.0.1
chromadb==0.5.5
rank-bm25==0.2.2

# PDF & parsing
pymupdf==1.24.9
unstructured==0.14.6

# Data ingestion
arxiv==2.1.3
semanticscholar==0.8.1

# Reranking (optional)
transformers==4.44.2
torch>=2.2

# Graph
py2neo==2021.2.4

# Utils
tqdm==4.66.5
requests==2.32.3
